% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjust_beta}
\alias{adjust_beta}
\title{Keep proportions inside the open interval (0, 1)}
\usage{
adjust_beta(x, epsilon = 0.001)
}
\arguments{
\item{x}{Numeric vector.}

\item{epsilon}{Small boundary value.}
}
\value{
A numeric vector with boundary values adjusted.
}
\description{
Keep proportions inside the open interval (0, 1)
}
