% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-models.R
\name{compare_models}
\alias{compare_models}
\title{Compare a candidate set using corrected Akaike information criterion}
\usage{
compare_models(models, competitive_delta = 2)
}
\arguments{
\item{models}{Named list of fitted models.}

\item{competitive_delta}{Maximum delta AICc defining competitive models.}
}
\value{
Model-selection data frame.
}
\description{
Compare a candidate set using corrected Akaike information criterion
}
