% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnose_models}
\alias{diagnose_models}
\title{Run DHARMa diagnostics for selected models}
\usage{
diagnose_models(object, nsim = 1000, seed = NULL)
}
\arguments{
\item{object}{An `ecoglmm_fit` object.}

\item{nsim}{Number of simulations.}

\item{seed}{Optional random seed.}
}
\value{
List containing a summary table and simulated residual objects.
}
\description{
Run DHARMa diagnostics for selected models
}
