% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_coefficients}
\alias{plot_coefficients}
\title{Plot coefficients and confidence intervals}
\usage{
plot_coefficients(object, include_intercept = FALSE)
}
\arguments{
\item{object}{An `ecoglmm_fit` object.}

\item{include_intercept}{Include intercept terms.}
}
\value{
A faceted ggplot object.
}
\description{
Plot coefficients and confidence intervals
}
