% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_effect}
\alias{plot_effect}
\title{Plot the marginal effect of a selected model}
\usage{
plot_effect(object, response, show_observed = TRUE)
}
\arguments{
\item{object}{An `ecoglmm_fit` object.}

\item{response}{Response name.}

\item{show_observed}{Add observed values.}
}
\value{
A ggplot object.
}
\description{
Plot the marginal effect of a selected model
}
