% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_selection}
\alias{plot_selection}
\title{Plot model-selection results}
\usage{
plot_selection(object, response, metric = c("delta", "weight"))
}
\arguments{
\item{object}{An `ecoglmm_fit` object.}

\item{response}{Response name.}

\item{metric}{Either `delta` or `weight`.}
}
\value{
A ggplot object.
}
\description{
Plot model-selection results
}
