% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-models.R
\name{test_interactions}
\alias{test_interactions}
\title{Likelihood-ratio tests for additive versus interaction models}
\usage{
test_interactions(models, predictors)
}
\arguments{
\item{models}{Named candidate-model list.}

\item{predictors}{Predictor names.}
}
\value{
Data frame of likelihood-ratio tests.
}
\description{
Likelihood-ratio tests for additive versus interaction models
}
